requirements.txt
flask==2.3.3
scikit-learn==1.3.2
scikit-surprise==1.1.3
pandas==1.5.3
numpy<2  # Important: this avoids compatibility issues with scikit-surprise
